Move hell to util
