+++ util/hell/cmd/hell/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/Masterminds/glide/msg"
 	"github.com/Masterminds/glide/path"
 	"github.com/Masterminds/glide/util"
-	"github.com/eris-ltd/eris-db/hell"
+	"github.com/eris-ltd/eris-db/util/hell"
 	"github.com/spf13/cobra"
 	"github.com/Masterminds/glide/action"
 	"github.com/Masterminds/glide/repo"
